+++ cli/rewatch.js
@@ -3,8 +3,10 @@
 // @ts-check
 
 import * as child_process from "node:child_process";
-import { rewatch_exe } from "./common/bins.js";
+import { rewatch_exe, bsc_exe } from "./common/bins.js";
 
 const args = process.argv.slice(2);
 
-child_process.spawnSync(rewatch_exe, args, { stdio: "inherit" });
+child_process.spawnSync(rewatch_exe, [...args, "--bsc-path", bsc_exe], {
+  stdio: "inherit",
+});